Move ground test to own package
