+++ common/src/test/java/net/theevilreaper/tamias/common/ground/GroundDataRegistryTest.java
@@ -1,8 +1,6 @@
-package net.theevilreaper.tamias.common.area.ground;
+package net.theevilreaper.tamias.common.ground;
 
 import net.minestom.server.instance.block.Block;
-import net.theevilreaper.tamias.common.ground.GroundData;
-import net.theevilreaper.tamias.common.ground.GroundDataRegistry;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
