+++ common/src/test/java/net/theevilreaper/tamias/common/ground/GroundDataTest.java
@@ -1,7 +1,6 @@
-package net.theevilreaper.tamias.common.area;
+package net.theevilreaper.tamias.common.ground;
 
 import net.minestom.server.instance.block.Block;
-import net.theevilreaper.tamias.common.ground.GroundData;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
